tileset refactor, menu animation

+++ src/globals.rs
@@ -1,2 +1,2 @@
-pub const PIXEL_PER_DOT: i32 = 64;
-pub const GROUND_LEVEL: i32 = 5;
+pub const PIXEL_PER_DOT: f64 = 64.0;
+pub const GROUND_LEVEL: f64 = 5.0;
